Add context provider to the App.js
